Infrastructure to to store and process MCMC results

The RexParams infrastructure is used to store and process MCMC results for Rex model from Crook et al 2024.

chains(object)

# S4 method for RexParams
show(object)

# S4 method for RexChain
show(object)

# S4 method for RexSummary
show(object)

# S4 method for RexChains
length(x)

# S4 method for RexParams
length(x)

posteriorEstimates(object)

# S4 method for RexSummary
posteriorEstimates(object)

Rex.quantiles(object)

# S4 method for RexSummary
Rex.quantiles(object)

Rex.resolution(object)

# S4 method for RexSummary
Rex.resolution(object)

Rex.peptideError(object)

# S4 method for RexSummary
Rex.peptideError(object)

Rex.globals(object)

# S4 method for RexSummary
Rex.globals(object)

# S4 method for RexParams
posteriorEstimates(object)

# S4 method for RexParams
Rex.quantiles(object)

# S4 method for RexParams
Rex.resolution(object)

# S4 method for RexParams
Rex.peptideError(object)

# S4 method for RexParams
Rex.globals(object)

# S4 method for RexChains,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for RexChains,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for RexParams,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for RexChains
show(object)

# S4 method for RexDifferential
show(object)

Rex.predictionError(object)

Rex.estimates(object)

Rex.probs(object)

Rex.eFDR(object)

# S4 method for RexDifferential
Rex.predictionError(object)

# S4 method for RexDifferential
Rex.estimates(object)

# S4 method for RexDifferential
Rex.probs(object)

# S4 method for RexDifferential
Rex.eFDR(object)

Arguments

object

An instance of appropriate class.

x

Object to be subset.

i

An integer(). Should be of length 1 for [[.

j

Missing.

drop

Missing.

Value

An object of class RexParams which stores the main results for the analysis when using Rex

Details

Objects of the RexParams class are created with the Rex() function These objects store the priors for the model and the results of the MCMC chains, which themselves are stored as an instance of class RexChains and can be accessed with the chains() function. A summary of the RexChains (or class RexSummary) can be further computed with the RexProcess function.

see the RexMS vignette for examples

Slots

chains

list() containing the individual full MCMC chain results in an RexChains instance. Each element must be a valid RexChain instance.

posteriorEstimates

A DataFrame documenting the posteriors in an RexSummary instance

diagnostics

A matrix of dimensions 1 by 2 containing the RexSummary diagnostics.

Rex.quantiles

A DataFrame of with R rows indicated quantiles for key parameters in the Rex model

Rex.globals

A DataFrame of with R rows indicated global parameters at the moment the standard deviation sigma and the model likelihoods

Rex.resolution

A DataFrame of with R rows indicated resolution metrics

Rex.peptideError

A DataFrame indicated peptide error

method

A character() storing the Rex method name

priors

A list() with the priors for the parameters

seed

An integer() with the random number generation seed.

summary

Object of class RexSummary the summarised MCMC results available in the RexParams instance.

chains

Object of class RexChains containing the full MCMC results in the RexParams instance

interval

A numeric() with the interval for the residues cover

dataset

character indicating the dataset

blong

matrix of dimensions R by numIter containing the blong values for the residues

pilong

matrix of dimensions R by numIter containing the pilong values for the residues

qlong

matrix of dimensions R by numIter containing the qlong values for the residues

dlong

matrix of dimensions R by numIter containing the dlong values for the residues

loglikelihood

numeric of length numIter containing the loglikelihood values

phi

numeric of length 1 containing the phi values for the model

UptakeGuess

list each with a matrix of dimensions R by numTimepoints containing the UptakeGuess values for the residues

numBreak

numeric of length numIter containing the number of breakpoints

Sigma

numeric of length numIter containing the Sigma values

R

integer of length 1 containing the number of residues

timepoints

numeric of length numTimepoints containing the timepoints

numIter

integer of length 1 containing the number of mcmc iterations

numPeptides

integer of length 1 containing the number of peptides

numTimepoints

integer of length 1 containing the number of timepoints

dataset

character indicating the dataset

Rex.predictionError

DFrame containing the prediction error

Rex.estimates

DFrame containing the estimates ARE, TRE, signed ARE

Rex.probs

DFrame containing the probabilities and total probabilities

Rex.eFDR

DFrame containing the eFDR