Function that compuate the maximum uptakes for all sequences in the dataset

Function that compuate the maximum uptakes for all sequences in the dataset

maxUptakes(res)

Arguments

res

An object of class DataFrame containing the results of the hdx-ms experiment

Value

Returns a numeric vector of maximum uptakes

Examples

data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

maxUptakes(res = BRD4_apo)
#>    [1] 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 18 18 18 18 18 18
#>   [25] 18 18 18 18 18 18 18 18 18 18 18 18 12 12 12 12 12 12 12 12 12 12 12 12
#>   [49] 12 12 12 12 12 12 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#>   [73] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 15 15 15 15 15 15
#>   [97] 15 15 15 15 15 15 15 15 15 15 15 15 17 17 17 17 17 17 17 17 17 17 17 17
#>  [121] 17 17 17 17 17 17  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#>  [145]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  8  8  8  8  8  8
#>  [169]  8  8  8  8  8  8  8  8  8  8  8  8 16 16 16 16 16 16 16 16 16 16 16 16
#>  [193] 16 16 16 16 16 16 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19
#>  [217] 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13  7  7  7  7  7  7
#>  [241]  7  7  7  7  7  7  7  7  7  7  7  7 10 10 10 10 10 10 10 10 10 10 10 10
#>  [265] 10 10 10 10 10 10  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#>  [289]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  4  4  4  4  4  4
#>  [313]  4  4  4  4  4  4  4  4  4  4  4  4 11 11 11 11 11 11 11 11 11 11 11 11
#>  [337] 11 11 11 11 11 11 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14
#>  [361] 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14  5  5  5  5  5  5
#>  [385]  5  5  5  5  5  5  5  5  5  5  5  5  8  8  8  8  8  8  8  8  8  8  8  8
#>  [409]  8  8  8  8  8  8 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11
#>  [433] 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22 22
#>  [457] 22 22 22 22 22 22 22 22 22 22 22 22  7  7  7  7  7  7  7  7  7  7  7  7
#>  [481]  7  7  7  7  7  7 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11 11
#>  [505] 17 17 17 17 17 17 17 17 17 17 17 17 17 17 17 17 17 17  4  4  4  4  4  4
#>  [529]  4  4  4  4  4  4  4  4  4  4  4  4  7  7  7  7  7  7  7  7  7  7  7  7
#>  [553]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#>  [577]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9 10 10 10 10 10 10
#>  [601] 10 10 10 10 10 10 10 10 10 10 10 10 12 12 12 12 12 12 12 12 12 12 12 12
#>  [625] 12 12 12 12 12 12 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13
#>  [649] 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 20 20 20 20 20 20
#>  [673] 20 20 20 20 20 20 20 20 20 20 20 20 21 21 21 21 21 21 21 21 21 21 21 21
#>  [697] 21 21 21 21 21 21  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#>  [721]  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  9  9  9  9  9  9
#>  [745]  9  9  9  9  9  9  9  9  9  9  9  9 10 10 10 10 10 10 10 10 10 10 10 10
#>  [769] 10 10 10 10 10 10  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9
#>  [793] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10  5  5  5  5  5  5
#>  [817]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#>  [841]  5  5  5  5  5  5  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#>  [865]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  6  6  6  6  6  6
#>  [889]  6  6  6  6  6  6  6  6  6  6  6  6 10 10 10 10 10 10 10 10 10 10 10 10
#>  [913] 10 10 10 10 10 10  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8
#>  [937]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#>  [961]  6  6  6  6  6  6  6  6  6  6  6  6  5  5  5  5  5  5  5  5  5  5  5  5
#>  [985]  5  5  5  5  5  5  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [1009]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  5  5  5  5  5  5
#> [1033]  5  5  5  5  5  5  5  5  5  5  5  5 13 13 13 13 13 13 13 13 13 13 13 13
#> [1057] 13 13 13 13 13 13 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12
#> [1081]  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9  9
#> [1105]  9  9  9  9  9  9  9  9  9  9  9  9  6  6  6  6  6  6  6  6  6  6  6  6
#> [1129]  6  6  6  6  6  6  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5
#> [1153] 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13  9  9  9  9  9  9
#> [1177]  9  9  9  9  9  9  9  9  9  9  9  9  4  4  4  4  4  4  4  4  4  4  4  4
#> [1201]  4  4  4  4  4  4  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [1225]  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  8  7  7  7  7  7  7
#> [1249]  7  7  7  7  7  7  7  7  7  7  7  7  9  9  9  9  9  9  9  9  9  9  9  9
#> [1273]  9  9  9  9  9  9  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [1297]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  8  8  8  8  8  8
#> [1321]  8  8  8  8  8  8  8  8  8  8  8  8 16 16 16 16 16 16 16 16 16 16 16 16
#> [1345] 16 16 16 16 16 16  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [1369]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7
#> [1393]  7  7  7  7  7  7  7  7  7  7  7  7 11 11 11 11 11 11 11 11 11 11 11 11
#> [1417] 11 11 11 11 11 11 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13 13
#> [1441] 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 17 17 17 17 17 17
#> [1465] 17 17 17 17 17 17 17 17 17 17 17 17  5  5  5  5  5  5  5  5  5  5  5  5
#> [1489]  5  5  5  5  5  5  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6
#> [1513]  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5  5 16 16 16 16 16 16
#> [1537] 16 16 16 16 16 16 16 16 16 16 16 16 17 17 17 17 17 17 17 17 17 17 17 17
#> [1561] 17 17 17 17 17 17 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16
#> [1585] 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 13 13 13 13 13 13
#> [1609] 13 13 13 13 13 13 13 13 13 13 13 13  6  6  6  6  6  6  6  6  6  6  6  6
#> [1633]  6  6  6  6  6  6 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12
#> [1657] 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 16 19 19 19 19 19 19
#> [1681] 19 19 19 19 19 19 19 19 19 19 19 19  9  9  9  9  9  9  9  9  9  9  9  9
#> [1705]  9  9  9  9  9  9 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21
#> [1729] 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21  8  8  8  8  8  8
#> [1753]  8  8  8  8  8  8  8  8  8  8  8  8 12 12 12 12 12 12 12 12 12 12 12 12
#> [1777] 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12 12
#> [1801]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7 11 11 11 11 11 11
#> [1825] 11 11 11 11 11 11 11 11 11 11 11 11 14 14 14 14 14 14 14 14 14 14 14 14
#> [1849] 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14 14
#> [1873] 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19 19
#> [1897] 19 19 19 19 19 19 19 19 19 19 19 19  5  5  5  5  5  5  5  5  5  5  5  5
#> [1921]  5  5  5  5  5  5 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#> [1945]  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6  6 10 10 10 10 10 10
#> [1969] 10 10 10 10 10 10 10 10 10 10 10 10  9  9  9  9  9  9  9  9  9  9  9  9
#> [1993]  9  9  9  9  9  9 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> [2017]  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4