Function that prepares exchange indexes for a given sequence

Function that prepares exchange indexes for a given sequence

prepareIndexes(res)

Arguments

res

An object of class DataFrame containing the results of the hdx-ms experiment

Value

Returns exchange indexes as a list

Examples

data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

prepareIndexes(res = BRD4_apo)
#> [[1]]
#>  [1]  3  4  5  6  8  9 10 11 12 13 14 16 17 18 19 20 21 22 23 24 25 26
#> 
#> [[2]]
#>  [1]  8  9 10 11 12 13 14 16 17 18 19 20 21 22 23 24 25 26
#> 
#> [[3]]
#>  [1] 14 16 17 18 19 20 21 22 23 24 25 26
#> 
#> [[4]]
#>  [1] 17 18 19 20 21 22 23 24 25 26
#> 
#> [[5]]
#>  [1] 29 30 31 32 33 34 36 38 39 40
#> 
#> [[6]]
#>  [1] 43 44 49 50 51 52 54 55 57 58 59 60 61 62 63
#> 
#> [[7]]
#>  [1] 43 44 49 50 51 52 54 55 57 58 59 60 61 62 63 64 65
#> 
#> [[8]]
#> [1] 68 69 70 71 72
#> 
#> [[9]]
#> [1] 70 71 72 73 74
#> 
#> [[10]]
#> [1] 71 72 73 74 75 76 77 78
#> 
#> [[11]]
#>  [1] 71 72 73 74 75 76 77 78 79 80 81 83 84 85 87 88
#> 
#> [[12]]
#>  [1] 71 72 73 74 75 76 77 78 79 80 81 83 84 85 87 88 89 90 91
#> 
#> [[13]]
#>  [1] 77 78 79 80 81 83 84 85 87 88 89 90 91
#> 
#> [[14]]
#> [1] 80 81 83 84 85 87 88
#> 
#> [[15]]
#>  [1] 80 81 83 84 85 87 88 89 90 91
#> 
#> [[16]]
#> [1] 81 83 84 85 87 88
#> 
#> [[17]]
#> [1] 81 83 84 85 87 88 89 90 91
#> 
#> [[18]]
#> [1] 84 85 87 88
#> 
#> [[19]]
#>  [1]  91  92  93  94  96  97  98  99 100 101 102
#> 
#> [[20]]
#>  [1]  91  92  93  94  96  97  98  99 100 101 102 103 105 106
#> 
#> [[21]]
#> [1] 94 96 97 98 99
#> 
#> [[22]]
#> [1]  94  96  97  98  99 100 101 102
#> 
#> [[23]]
#>  [1]  94  96  97  98  99 100 101 102 103 105 106
#> 
#> [[24]]
#>  [1]  94  96  97  98  99 100 101 102 103 105 106 107 108 109 110 111 112 113 114
#> [20] 115 116 117
#> 
#> [[25]]
#> [1]  96  97  98  99 100 101 102
#> 
#> [[26]]
#>  [1] 100 101 102 103 105 106 107 108 109 110 111
#> 
#> [[27]]
#>  [1] 100 101 102 103 105 106 107 108 109 110 111 112 113 114 115 116 117
#> 
#> [[28]]
#> [1] 102 103 105 106
#> 
#> [[29]]
#> [1] 105 106 107 108 109 110 111
#> 
#> [[30]]
#> [1] 105 106 107 108 109 110 111 112 113
#> 
#> [[31]]
#>  [1] 105 106 107 108 109 110 111 112 113 114
#> 
#> [[32]]
#>  [1] 105 106 107 108 109 110 111 112 113 114 115 116
#> 
#> [[33]]
#>  [1] 105 106 107 108 109 110 111 112 113 114 115 116 117
#> 
#> [[34]]
#>  [1] 105 106 107 108 109 110 111 112 113 114 115 116 117 118
#> 
#> [[35]]
#>  [1] 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123
#> [20] 124
#> 
#> [[36]]
#>  [1] 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123
#> [20] 124 125
#> 
#> [[37]]
#> [1] 109 110 111 112 113
#> 
#> [[38]]
#> [1] 109 110 111 112 113 114 115 116
#> 
#> [[39]]
#> [1] 109 110 111 112 113 114 115 116 117
#> 
#> [[40]]
#>  [1] 109 110 111 112 113 114 115 116 117 118
#> 
#> [[41]]
#> [1] 116 117 118 119 120 121 122 123 124
#> 
#> [[42]]
#>  [1] 116 117 118 119 120 121 122 123 124 125
#> 
#> [[43]]
#> [1] 120 121 122 123 124
#> 
#> [[44]]
#> [1] 127 128 129 130 131
#> 
#> [[45]]
#> [1] 127 128 129 130 131 132
#> 
#> [[46]]
#> [1] 128 129 130 131 132
#> 
#> [[47]]
#> [1] 132 133 134 135 136 137
#> 
#> [[48]]
#>  [1] 138 139 140 141 143 144 145 146 147 148
#> 
#> [[49]]
#> [1] 140 141 143 144 145 146 147 148
#> 
#> [[50]]
#> [1] 146 147 148 149 150 151
#> 
#> [[51]]
#> [1] 150 151 152 153 154 155
#> 
#> [[52]]
#> [1] 151 152 153 154 155
#> 
#> [[53]]
#> [1] 151 152 153 154 155 156
#> 
#> [[54]]
#> [1] 152 153 154 155 156
#> 
#> [[55]]
#>  [1] 159 160 161 162 163 164 166 167 168 169 170 171 172
#> 
#> [[56]]
#>  [1] 160 161 162 163 164 166 167 168 169 170 171 172
#> 
#> [[57]]
#> [1] 163 164 166 167 168 169 170 171 172
#> 
#> [[58]]
#> [1] 164 166 167 168 169 170 171 172 173
#> 
#> [[59]]
#> [1] 167 168 169 170 171 172
#> 
#> [[60]]
#> [1] 367 368 369 370 371
#> 
#> [[61]]
#>  [1] 367 368 369 370 371 372 373 374 376 377 378 380 381
#> 
#> [[62]]
#> [1] 371 372 373 374 376 377 378 380 381
#> 
#> [[63]]
#> [1] 374 376 377 378
#> 
#> [[64]]
#> [1] 374 376 377 378 380 381
#> 
#> [[65]]
#> [1] 374 376 377 378 380 381 382 383
#> 
#> [[66]]
#> [1] 384 385 386 387 388 389 390
#> 
#> [[67]]
#> [1] 384 385 386 387 388 389 390 391 392
#> 
#> [[68]]
#> [1] 385 386 387 388 389 390
#> 
#> [[69]]
#> [1] 386 387 388 389 390
#> 
#> [[70]]
#> [1] 391 392 393 394 395 396 398 399
#> 
#> [[71]]
#>  [1] 391 392 393 394 395 396 398 399 400 401 402 403 404 405 406 407
#> 
#> [[72]]
#> [1] 392 393 394 395 396 398 399
#> 
#> [[73]]
#> [1] 393 394 395 396 398 399
#> 
#> [[74]]
#> [1] 393 394 395 396 398 399 400
#> 
#> [[75]]
#>  [1] 393 394 395 396 398 399 400 401 402 403 404
#> 
#> [[76]]
#>  [1] 393 394 395 396 398 399 400 401 402 403 404 405 406
#> 
#> [[77]]
#>  [1] 393 394 395 396 398 399 400 401 402 403 404 405 406 407
#> 
#> [[78]]
#>  [1] 393 394 395 396 398 399 400 401 402 403 404 405 406 407 408 409 410
#> 
#> [[79]]
#> [1] 402 403 404 405 406
#> 
#> [[80]]
#> [1] 402 403 404 405 406 407
#> 
#> [[81]]
#> [1] 403 404 405 406 407
#> 
#> [[82]]
#>  [1] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424
#> 
#> [[83]]
#>  [1] 409 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425
#> 
#> [[84]]
#>  [1] 410 411 412 413 414 415 416 417 418 419 420 421 422 423 424 425
#> 
#> [[85]]
#>  [1] 413 414 415 416 417 418 419 420 421 422 423 424 425
#> 
#> [[86]]
#> [1] 420 421 422 423 424 425
#> 
#> [[87]]
#>  [1] 428 429 430 431 432 433 436 437 438 439 440 441
#> 
#> [[88]]
#>  [1] 428 429 430 431 432 433 436 437 438 439 440 441 442 443 444 445
#> 
#> [[89]]
#>  [1] 428 429 430 431 432 433 436 437 438 439 440 441 442 443 444 445 446 447 448
#> 
#> [[90]]
#> [1] 431 432 433 436 437 438 439 440 441
#> 
#> [[91]]
#>  [1] 431 432 433 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451
#> [20] 452 453
#> 
#> [[92]]
#> [1] 432 433 436 437 438 439 440 441
#> 
#> [[93]]
#>  [1] 432 433 436 437 438 439 440 441 442 443 444 445
#> 
#> [[94]]
#> [1] 433 436 437 438 439 440 441
#> 
#> [[95]]
#>  [1] 433 436 437 438 439 440 441 442 443 444 445
#> 
#> [[96]]
#>  [1] 433 436 437 438 439 440 441 442 443 444 445 446 447 448
#> 
#> [[97]]
#>  [1] 433 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450 451 452 453
#> 
#> [[98]]
#> [1] 444 445 446 447 448
#> 
#> [[99]]
#>  [1] 444 445 446 447 448 449 450 451 452 453
#> 
#> [[100]]
#> [1] 448 449 450 451 452 453
#> 
#> [[101]]
#>  [1] 455 456 457 459 460 462 463 465 466 467
#> 
#> [[102]]
#> [1] 456 457 459 460 462 463 465 466 467
#> 
#> [[103]]
#>  [1] 456 457 459 460 462 463 465 466 467 468 469 470 472 473 477
#> 
#> [[104]]
#> [1] 470 472 473 477
#>